move popup logic to popup.js
